+++ player.h
@@ -8,7 +8,7 @@ namespace Player
 extern uint8_t hp;
 extern uint16_t score;
   
-void init(int16_t x, int16_t y);
+void init(int16_t x, int8_t y);
 void update();
 void draw();
 }
